fix: change langSwitcher
